+++ app/src/main/java/com/ratbyansa/moviedb/data/local/entity/GenreEntity.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "genres")
 data class GenreEntity(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     val name: String
 )
